Safari-609.1.11.1/LayoutTests/fast/speechsynthesis/speech-synthesis-speak-empty-string-expected.txt

This tests that an empty string sent to the synthesizer still generates a callback.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS speechSynthesis.speaking is false
PASS event.elapsedTime > 0 is true
PASS speechSynthesis.speaking is false
PASS successfullyParsed is true

TEST COMPLETE
